Renamed a few methods + handle not safe moves
Renamed a few methods, to make it easier to understand the flow.
In getMoveDirection(), if no "good" moves are left, we are now trying
to find a move, that is "safe". Meaning not directly colliding with a
current snake position (NOT looking at where a snake MIGHT BE in the
next move)